refactor(frontend): 拆分 LightField 为舞台/装饰层/粒子形态模块
- Bot Review Blocking：单文件 God component 拆为 stage/decor/shapes/config 四模块，组件保留薄壳

- 魔法数（tone 阈值、PARALLAX、HALO_LERP、GLOW_RADIUS、LINK_MAX）提为具名常量

- 顺带修复 drawLinks O(n²)：近层连线候选 cap 36 个，判定数约 1770→630/帧
